requirements.txt
absl-py==2.4.0
anyio==4.12.1
backports.zstd @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_backports.zstd_1767044984/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1764016952863/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1767500808759/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
grpcio==1.76.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hf-xet==1.2.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.7
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
Markdown==3.10.1
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
mxnet==1.5.0
networkx==3.4.2
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1718615061894/work/dist/numpy-2.0.0-cp310-cp310-linux_x86_64.whl#sha256=91843c84a6c7fb01afed16468f942dfe32b48184a79039847a16b4459e5735c4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.13.0.90
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.3.3
pillow==12.1.0
protobuf==4.25.8
pyparsing==3.3.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_requests_1766926974/work
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.2
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.67.2
transformers==5.0.0
triton==3.0.0
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1767817748113/work
Werkzeug==3.1.5
